requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.1.0
pymongo>=4.6.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiohttp>=3.9.0
async-timeout>=4.0.0